feat: added database connection with Prisma client
- Create singleton Prisma client to prevent multiple instances
- Configure logging based on environment
- Set up global reference for development hot reloading
- Add initial database migration files

thought behind NODE_ENV is, that I can decide what to show based on the type of instance (prod or dev). example in another project of mine: in dev environment I have created a page, with multiple tests, that help automate tests, that would otherwise take too long. NODE_ENV is being set in .env (or in the environment variables if prod is run on a host like Coolify.

migration_lock.toml and migration.sql are both files that were being generated when ```npx prisma migrate dev``` is being run. (they define the actual database that will be created in the docker instance)